refactored avalanche dependency
