+++ avalanche_rl/training/strategies/__init__.py
@@ -1,4 +1,5 @@
-from .vectorized_env import VectorizedEnvironment
 from .rl_base_strategy import RLBaseStrategy
 from .dqn import DQNStrategy
 from .actor_critic import A2CStrategy
+
+__all__ = ['RLBaseStrategy', 'A2CStrategy', 'DQNStrategy']
